[REFACTOR] : 파일 분리
