+++ package.json
@@ -5,7 +5,7 @@
   "description": "강원도 내 반려동물 동반 가능 시설 api 활용 홈페이지",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "vitest",
     "format": "prettier --write --cache ."
   },
   "keywords": [],
@@ -17,7 +17,8 @@
     "eslint-config-prettier": "^9.0.0",
     "eslint-plugin-import": "^2.29.0",
     "eslint-plugin-prettier": "^5.0.1",
-    "prettier": "^3.1.0"
+    "prettier": "^3.1.0",
+    "vitest": "^1.0.1"
   },
   "dependencies": {
     "express": "^4.18.2",
