+++ public/style.css
@@ -18,7 +18,7 @@
 }
 
 #table {
-    visibility: hidden;
+    visibility:hidden;
     margin: auto;
     margin-bottom: 100px;
     border: black;
